Experimental global edit keys Action Handler
Bind Cell Editors and Text Editors to Global Key Action Handlers so that Cut/Copy/Paste/Delete global actions are enabled when editing text in cell and text editors.

PROBLEM: The Cut/Copy/Paste/Delete toolbar items doesn't work. See bug https://bugs.eclipse.org/bugs/show_bug.cgi?id=321045 Although it is possible to copy the CellEditorActionHandler class and change the events to FocusIn and FocusOut.

There just seems too much work and too many workarounds for the result.

Even without this patch, at least the shortcut keys in the cell and text editors (Ctrl-A, Ctrl-C, Ctrl-X, Ctrl-V, Del) work, so it's not the end of the world.

Perhaps the best thing to do is to set the Global Action Handlers to null for the Cut/Copy/Paste/Delete actions when a cell editor is activated. This way any underlying selected objects in editors or the tree are not copied, cut or pasted.